+++ driver/eeprom.h
@@ -23,7 +23,7 @@
 #include <stdint.h>
 #include <stdbool.h>
 
-void EEPROM_ReadBuffer(uint16_t Address, void *pBuffer, uint16_t Size);
+void EEPROM_ReadBuffer(uint16_t Address, void *pBuffer, uint8_t Size);
 void EEPROM_WriteBuffer(uint16_t Address, const void *pBuffer);
 
 #endif
